# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PhotosDen
  version: 1.0.0
  description: Secure, serverless personal photo storage on AWS.
servers:
  - url: https://g2l2bkihl6.execute-api.ap-south-1.amazonaws.com/dev
    description: Development Environment
# Security is now handled in Lambda for better CORS control
# security:
#   - BearerAuth: []
paths:
...
  /shares/{id}:
    delete:
      summary: Archive share history
      tags: [Shares]
      security:
        - BearerAuth: []
      x-amazon-apigateway-integration:
        httpMethod: POST
        type: aws_proxy
        uri: ${lambda_invoke_arn}
        payloadFormatVersion: "2.0"
        responses:
          default:
            statusCode: "204"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Share history archived
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/X-Correlation-Id"
...
  /share/{token}:
    get:
      summary: Public share access
      x-amazon-apigateway-integration:
        httpMethod: POST
        type: aws_proxy
        uri: ${lambda_invoke_arn}
        payloadFormatVersion: "2.0"
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared content metadata
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/X-Correlation-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicShareResponse"
        "404":
          description: Link not found, expired, or target deleted
        "410":
          description: Link target is in trash
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito User Pool Access Token
      x-amazon-apigateway-authorizer:
        type: "cognito_user_pools"
        providerArns:
          - "${cognito_user_pool_arn}"
      x-amazon-apigateway-authtype: "cognito_user_pools"
  headers:
    X-Correlation-Id:
      description: Unique ID for request tracing
      schema:
        type: string
...
  schemas:
...
    PublicShareResponse:
      type: object
      required: [targetMetadata]
      properties:
        targetMetadata:
          type: object
          properties:
            title:
              type: string
            itemCount:
              type: integer
        items:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              assetId:
                type: string
...